Put in the working app please see description
Each main page is one class which sets up the main windows/frames and includes the functions for navigating. There are now 2 python packages called AdminSubpages and VolunteerSubpages; these contain files of functions for each of the frames within the hompages for both admin and volunteers. Everything currently in Volunteer Profile Page is split between volunteer_log_in, volunteer_main, and the relevant package. This means we do not have to deal with lengthy code and we can work on different frames more clearly; especially when it comes to the end of the project and GUI needs to look decent.

NOTE: I have not deleted any of the previous files. We still have resource allocation (which will be integrated into the admin package) and the Volunteer Profile Page.py. I have put any redundant python files into a folder called LearningResourcesAndOldFiles as these do not contribute to the project. I would advise analysing the method and working directly on each folder within the packages whilst referencing volunteer/admin_main.py as integrating into the app after the code is written separately can be quite a challenge.