Resource allocation - incorporating into main
Seb and I have started incorporating this into the admin class.

I made a new file called panda data handler. This could be a new class that defines functions for handling the data inputs and converting it to the pandas. We could reuse this throughout the whole programme. We can discuss this in detail.